# contracts/futures/v1/internal-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Futures worker execution",
    "version": "F-1.0"
  },
  "paths": {
    "/internal/futures/v1/grants": {
      "post": {
        "operationId": "issueFuturesGrant",
        "security": [
          {
            "serviceIdentity": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Short single-use grant"
          },
          "default": {
            "description": "Denied"
          }
        }
      }
    },
    "/internal/futures/v1/tools": {
      "post": {
        "operationId": "callFuturesTool",
        "security": [
          {
            "serviceIdentity": [],
            "taskGrant": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool result"
          },
          "default": {
            "description": "Denied or unavailable"
          }
        }
      }
    },
    "/internal/futures/v1/model": {
      "post": {
        "operationId": "callFuturesModel",
        "security": [
          {
            "serviceIdentity": [],
            "taskGrant": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model result"
          },
          "default": {
            "description": "Denied or unavailable"
          }
        }
      }
    },
    "/internal/futures/v1/execute": {
      "post": {
        "operationId": "executeFutures",
        "security": [
          {
            "serviceIdentity": [],
            "taskGrant": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./types.schema.json#/$defs/Task"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Candidate report, not permission to publish",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./types.schema.json#/$defs/Report"
                }
              }
            }
          },
          "default": {
            "description": "Execution rejected or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./types.schema.json#/$defs/Error"
                }
              }
            }
          }
        }
      }
    },
    "/internal/futures/v1/observations": {
      "post": {
        "operationId": "ingestFuturesObservations",
        "security": [
          {
            "serviceIdentity": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationIngestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inserted normalized observations and condition evaluations"
          },
          "default": {
            "description": "Denied or invalid"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "GrantRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "domain": {
            "const": "futures"
          },
          "owner_id": {
            "type": "integer",
            "minimum": 1
          },
          "mode": {
            "const": "live"
          },
          "run_id": {
            "type": "string",
            "minLength": 1
          },
          "task_id": {
            "type": "string",
            "minLength": 1
          },
          "generation": {
            "type": "integer",
            "minimum": 1
          },
          "manifest_id": {
            "type": "string",
            "minLength": 1
          },
          "stage": {
            "enum": [
              "evidence",
              "support",
              "challenge",
              "verify",
              "repair",
              "execute"
            ]
          },
          "tool_name": {
            "enum": [
              "",
              "futures_get_observations",
              "futures_get_evidence",
              "futures_calculate"
            ]
          },
          "args_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        },
        "required": [
          "domain",
          "owner_id",
          "mode",
          "run_id",
          "task_id",
          "generation",
          "manifest_id",
          "stage",
          "tool_name",
          "args_hash"
        ]
      },
      "InternalScope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "domain": {
            "const": "futures"
          },
          "mode": {
            "const": "live"
          },
          "owner_id": {
            "type": "integer",
            "minimum": 1
          },
          "run_id": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "domain",
          "mode",
          "owner_id",
          "run_id"
        ]
      },
      "ToolRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/InternalScope"
          },
          "task_id": {
            "type": "string",
            "minLength": 1
          },
          "generation": {
            "type": "integer",
            "minimum": 1
          },
          "manifest_id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "enum": [
              "futures_get_observations",
              "futures_get_evidence",
              "futures_calculate"
            ]
          },
          "arguments": {
            "type": "object"
          }
        },
        "required": [
          "scope",
          "task_id",
          "generation",
          "manifest_id",
          "name",
          "arguments"
        ]
      },
      "ModelRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/InternalScope"
          },
          "task_id": {
            "type": "string",
            "minLength": 1
          },
          "generation": {
            "type": "integer",
            "minimum": 1
          },
          "manifest_id": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "const": "futures_model"
          },
          "arguments": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "stage": {
                "enum": [
                  "support",
                  "challenge",
                  "verify",
                  "repair"
                ]
              },
              "messages": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              },
              "max_output_tokens": {
                "type": "integer",
                "minimum": 1,
                "maximum": 4096
              }
            },
            "required": [
              "stage",
              "messages"
            ]
          }
        },
        "required": [
          "scope",
          "task_id",
          "generation",
          "manifest_id",
          "name",
          "arguments"
        ]
      },
      "ObservationIngestRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "owner_id": {
            "type": "integer",
            "minimum": 1
          },
          "mode": {
            "const": "live"
          },
          "records": {
            "type": "array",
            "minItems": 1,
            "maxItems": 500,
            "items": {
              "$ref": "./types.schema.json#/$defs/Record"
            }
          }
        },
        "required": [
          "owner_id",
          "mode",
          "records"
        ]
      }
    },
    "securitySchemes": {
      "serviceIdentity": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Futures-Service-Token"
      },
      "taskGrant": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}